Chaging timeout datatype to unsigned int
This allows a small compiler optimization on different
architectures.

See https://github.com/arduino-libraries/SD/pull/43#issuecomment-344604465

+++ src/utility/Sd2Card.cpp
@@ -248,7 +248,7 @@ uint8_t Sd2Card::init(uint8_t sckRateID, uint8_t chipSelectPin) {
   errorCode_ = inBlock_ = partialBlockRead_ = type_ = 0;
   chipSelectPin_ = chipSelectPin;
   // 16-bit init start time allows over a minute
-  uint16_t t0 = millis();
+  unsigned int t0 = millis();
   uint32_t arg;
 
   // set pin modes
@@ -288,7 +288,7 @@ uint8_t Sd2Card::init(uint8_t sckRateID, uint8_t chipSelectPin) {
 
   // command to go idle in SPI mode
   while ((status_ = cardCommand(CMD0, 0)) != R1_IDLE_STATE) {
-    uint16_t d = millis() - t0;
+    unsigned int d = millis() - t0;
     if (d > SD_INIT_TIMEOUT) {
       error(SD_CARD_ERROR_CMD0);
       goto fail;
@@ -311,7 +311,7 @@ uint8_t Sd2Card::init(uint8_t sckRateID, uint8_t chipSelectPin) {
 
   while ((status_ = cardAcmd(ACMD41, arg)) != R1_READY_STATE) {
     // check for timeout
-    uint16_t d = millis() - t0;
+    unsigned int d = millis() - t0;
     if (d > SD_INIT_TIMEOUT) {
       error(SD_CARD_ERROR_ACMD41);
       goto fail;
@@ -543,9 +543,9 @@ uint8_t Sd2Card::setSpiClock(uint32_t clock)
 #endif
 //------------------------------------------------------------------------------
 // wait for card to go not busy
-uint8_t Sd2Card::waitNotBusy(uint16_t timeoutMillis) {
-  uint16_t t0 = millis();
-  uint16_t d;
+uint8_t Sd2Card::waitNotBusy(unsigned int timeoutMillis) {
+  unsigned int t0 = millis();
+  unsigned int d;
   do {
     if (spiRec() == 0XFF) return true;
     d = millis() - t0;
@@ -556,9 +556,9 @@ uint8_t Sd2Card::waitNotBusy(uint16_t timeoutMillis) {
 //------------------------------------------------------------------------------
 /** Wait for start block token */
 uint8_t Sd2Card::waitStartBlock(void) {
-  uint16_t t0 = millis();
+  unsigned int t0 = millis();
   while ((status_ = spiRec()) == 0XFF) {
-    uint16_t d = millis() - t0;
+    unsigned int d = millis() - t0;
     if (d > SD_READ_TIMEOUT) {
       error(SD_CARD_ERROR_READ_TIMEOUT);
       goto fail;
